requirements.txt
pandas==1.3.3
streamlit==1.14.0
Pillow==8.3.2
requests==2.26.0